# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Parallel AI Coding Orchestrator API",
    "description": "\n## Week 2 MVP - Manager AI Integration API\n\nEnterprise-grade REST API for orchestrating and monitoring parallel AI coding workflows.\n\n### Key Features\n- **Supervisor API**: Monitor and control Claude Code worker instances\n- **Resource Management**: Hierarchical resource allocation with depth-based quotas\n- **Job Orchestration**: Manage job lifecycle with state machine validation\n- **Authentication**: JWT-based authentication with scope-based authorization\n\n### Architecture\n- **Database**: SQLite (development) / PostgreSQL (production)\n- **Authentication**: JWT Bearer tokens with scopes\n- **State Management**: Validated state machines for workers and jobs\n- **Testing**: 44 integration tests with 29% code coverage\n\n### Quick Start\n1. Generate test token: Use `/api/auth/token` endpoint\n2. Access protected endpoints with `Authorization: Bearer <token>` header\n3. Explore interactive docs at `/docs` or `/redoc`\n\n### Support\n- **GitHub**: [parallel-coding](https://github.com/your-org/parallel-coding)\n- **Documentation**: [Wiki](https://github.com/your-org/parallel-coding/wiki)\n",
    "contact": {
      "name": "Parallel AI Development Team",
      "email": "dev@parallel-ai.example.com"
    },
    "license": {
      "name": "MIT"
    },
    "version": "2.0.0-week2-mvp"
  },
  "paths": {
...
    "/api/jobs/submit": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Submit a new job",
        "description": "Create and submit a new job for hierarchical execution.\n\n    **Request Body:**\n    - `task_description`: Task to execute (required, 1-4096 chars)\n    - `worker_count`: Number of workers to allocate (1-1000)\n    - `depth`: Hierarchy depth level (0-1000, default: 0)\n    - `parent_job_id`: Parent job ID for hierarchical jobs (optional)\n\n    **Job Lifecycle:**\n    ```\n    SUBMITTED → PENDING → RUNNING → COMPLETED\n             → PENDING → RUNNING → FAILED\n             → PENDING → CANCELED\n    ```\n\n    **Required Scope:** `jobs:write`\n\n    **Example Request:**\n    ```json\n    {\n      \"task_description\": \"Implement user authentication\",\n      \"worker_count\": 3,\n      \"depth\": 0\n    }\n    ```\n\n    **Example Response:**\n    ```json\n    {\n      \"id\": \"j_a1b2c3d4e5f6\",\n      \"status\": \"PENDING\",\n      \"task_description\": \"Implement user authentication\",\n      \"worker_count\": 3,\n      \"depth\": 0,\n      \"parent_job_id\": null,\n      \"created_at\": \"2025-10-28T10:00:00Z\",\n      \"updated_at\": \"2025-10-28T10:00:01Z\"\n    }\n    ```",
        "operationId": "submit_job_api_jobs_submit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobSubmitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created job with PENDING status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/jobs/{job_id}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Job",
        "description": "Return job details by identifier.",
        "operationId": "get_job_api_jobs__job_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}/cancel": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Cancel Job",
        "description": "Cancel a job using the state machine.\n\nReturns the updated job representation. If cancellation is not allowed\nfrom the current state, returns ``400``.",
        "operationId": "cancel_job_api_jobs__job_id__cancel_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "List Jobs",
        "description": "List jobs with optional filters and pagination.",
        "operationId": "list_jobs_api_jobs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "depth",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Depth"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/JobStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "parent_job_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Parent Job Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobResponse"
                  },
                  "title": "Response List Jobs Api Jobs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JobResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "depth": {
            "type": "integer",
            "title": "Depth"
          },
          "worker_count": {
            "type": "integer",
            "title": "Worker Count"
          },
          "task_description": {
            "type": "string",
            "title": "Task Description"
          },
          "parent_job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Job Id"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "depth",
          "worker_count",
          "task_description",
          "parent_job_id",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "JobResponse",
        "description": "Serialized representation of a Job."
      },
      "JobStatus": {
        "type": "string",
        "enum": [
          "SUBMITTED",
          "PENDING",
          "RUNNING",
          "COMPLETED",
          "FAILED",
          "CANCELED"
        ],
        "title": "JobStatus",
        "description": "Job lifecycle status enumeration.\n\nStates:\n    SUBMITTED: Job received but not yet queued\n    PENDING: Job queued, waiting for resources\n    RUNNING: Job actively being processed\n    COMPLETED: Job finished successfully\n    FAILED: Job encountered fatal error\n    CANCELED: Job canceled by user"
      },
      "JobSubmitRequest": {
        "properties": {
          "task_description": {
            "type": "string",
            "maxLength": 4096,
            "minLength": 1,
            "title": "Task Description"
          },
          "worker_count": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 1.0,
            "title": "Worker Count"
          },
          "depth": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 0.0,
            "title": "Depth",
            "default": 0
          },
          "parent_job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Job Id"
          }
        },
        "type": "object",
        "required": [
          "task_description",
          "worker_count"
        ],
        "title": "JobSubmitRequest",
        "description": "Request body for submitting a new job."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}